Auto Train Code completed (Mostly)
Code modified to automate training. Results can be seen in  Decision Tree folder. Not configured is Max Leaf,Min Sample split, Min Leaf. With the default values for those hyper paramters not configured, the best score received is 59.55% with a max depth of 24, min impurity decrease of 0, max features set to "sqrt", and using the gini criterion. Still need to decide range of values to use in non-configured hyper parameters and what order to loop through them if order matters.