a few changes,same accuracy
I rearranged the order of loops with the categorical values first. Then they are ordered least to most range of values for parameters. After a period of 2 observations I found that maxLeaf resulted in 0% accuracy at value 2 and a minimum impurity decrease higher than 0.0 results in 0% accuracy. I proceeded to set these parameters to default values until further research can be completed. there is a printing error in the 4th picture with minLeaf because I accidently set the minLeaf key's value to the name of the variable,not the current minLeaf value. The minLeaf value that results in 59.55% accuracy is 1.