visualization
added visualization of the decision tree. In doing so I replaced the semi-colon with commas because it was annoying. I also realized that the data was excluding the alcohol column so it is now fixed. With the last, best hyper paramter configurations the accuracy dropped 2 points. Further testing is required to find new best paramter configurations. tree can be seen under data folder.